Many changes
- Implemented GUI display for fuel tanks
- Refactored code for printing details
- Added methods for getting status and status number to FuelTankStation

+++ Assignments/Assignments/GasStation/Customer.h
@@ -7,7 +7,7 @@ class Customer :
 
 private:
 	int pumpNumber;
-	int creditCard;
+	long long creditCard;
 	double requestedVolume;
 	int fuelGrade;
